refactor: rewrite cjs module to esm module
